+++ src/CaptureClient/include/CaptureClient/CaptureClient.h
@@ -42,7 +42,7 @@ class CaptureClient {
       orbit_base::ThreadPool* thread_pool, int32_t process_id,
       const orbit_client_data::ModuleManager& module_manager,
       absl::flat_hash_map<uint64_t, orbit_client_protos::FunctionInfo> selected_functions,
-      bool always_record_arguments, bool record_return_values,
+      bool record_arguments, bool record_return_values,
       orbit_client_data::TracepointInfoSet selected_tracepoints, double samples_per_second,
       uint16_t stack_dump_size, orbit_grpc_protos::UnwindingMethod unwinding_method,
       bool collect_scheduling_info, bool collect_thread_state, bool collect_gpu_jobs,
@@ -71,14 +71,11 @@ class CaptureClient {
 
   bool AbortCaptureAndWait(int64_t max_wait_ms);
 
-  [[nodiscard]] static orbit_grpc_protos::InstrumentedFunction::FunctionType
-  InstrumentedFunctionTypeFromOrbitType(orbit_client_protos::FunctionInfo::OrbitType orbit_type);
-
  private:
   ErrorMessageOr<CaptureListener::CaptureOutcome> CaptureSync(
       int32_t process_id, const orbit_client_data::ModuleManager& module_manager,
       const absl::flat_hash_map<uint64_t, orbit_client_protos::FunctionInfo>& selected_functions,
-      bool always_record_arguments, bool record_return_values,
+      bool record_arguments, bool record_return_values,
       const orbit_client_data::TracepointInfoSet& selected_tracepoints, double samples_per_second,
       uint16_t stack_dump_size, orbit_grpc_protos::UnwindingMethod unwinding_method,
       bool collect_scheduling_info, bool collect_thread_state, bool collect_gpu_jobs,
