+++ src/ClientData/ModuleManagerTest.cpp
@@ -247,101 +247,6 @@ TEST(ModuleManager, AddOrUpdateNotLoadedModules) {
   EXPECT_EQ(different_module->file_size(), different_file_size);
 }
 
-TEST(ModuleManager, GetOrbitFunctionsOfProcess) {
-  constexpr const char* kName = "name of module";
-  constexpr const char* kFilePath = "path/of/module";
-  constexpr const char* kBuildId = "build id example";
-  constexpr const char* kBuildIdChanged = "another build id";
-  constexpr uint64_t kFileSize = 300;
-  constexpr uint64_t kLoadBias = 0x400;
-  constexpr uint64_t kAddressStart = 100;
-  constexpr uint64_t kAddressEnd = 1000;
-
-  ModuleInfo module_info;
-  module_info.set_name(kName);
-  module_info.set_file_path(kFilePath);
-  module_info.set_file_size(kFileSize);
-  module_info.set_build_id(kBuildId);
-  module_info.set_load_bias(kLoadBias);
-  module_info.set_address_start(kAddressStart);
-  module_info.set_address_end(kAddressEnd);
-
-  ModuleManager module_manager;
-  EXPECT_TRUE(module_manager.AddOrUpdateModules({module_info}).empty());
-
-  ProcessInfo process_info;
-  ProcessData process{process_info};
-
-  {
-    // process does not have any modules loaded
-    std::vector<FunctionInfo> functions = module_manager.GetOrbitFunctionsOfProcess(process);
-    EXPECT_TRUE(functions.empty());
-  }
-
-  process.UpdateModuleInfos({module_info});
-  {
-    // process has module loaded, but module does not have symbols loaded
-    std::vector<FunctionInfo> functions = module_manager.GetOrbitFunctionsOfProcess(process);
-    EXPECT_TRUE(functions.empty());
-  }
-
-  ModuleSymbols module_symbols;
-  {
-    SymbolInfo* symbol_info = module_symbols.add_symbol_infos();
-    symbol_info->set_name("mangled_not_an_orbit_function");
-    symbol_info->set_demangled_name("not an orbit function");
-  }
-
-  const auto& orbit_name_to_orbit_type_map = function_utils::GetFunctionNameToOrbitTypeMap();
-  const std::string orbit_name = orbit_name_to_orbit_type_map.begin()->first;
-  const std::string orbit_name_mangled = absl::StrFormat("mangled_%s", orbit_name);
-
-  {
-    SymbolInfo* symbol_info = module_symbols.add_symbol_infos();
-    symbol_info->set_name(orbit_name_mangled);
-    symbol_info->set_demangled_name(orbit_name);
-    symbol_info->set_address(500);
-  }
-
-  ModuleData* module = module_manager.GetMutableModuleByPathAndBuildId(kFilePath, kBuildId);
-  module->AddSymbols(module_symbols);
-
-  {
-    // process has module loaded and module has functions loaded. One of the functions is an orbit
-    // function
-    std::vector<FunctionInfo> functions = module_manager.GetOrbitFunctionsOfProcess(process);
-    ASSERT_EQ(functions.size(), 1);
-
-    FunctionInfo& function{functions[0]};
-    EXPECT_EQ(function.name(), orbit_name_mangled);
-    EXPECT_EQ(function.pretty_name(), orbit_name);
-    EXPECT_EQ(function.address(), 500);
-  }
-
-  module_info.set_build_id(kBuildIdChanged);
-  // Add same module with another build_id
-  EXPECT_TRUE(module_manager.AddOrUpdateModules({module_info}).empty());
-  {
-    // We should still be able to lookup the function because old module did not go anywhere.
-    std::vector<FunctionInfo> functions = module_manager.GetOrbitFunctionsOfProcess(process);
-    ASSERT_EQ(functions.size(), 1);
-
-    FunctionInfo& function{functions[0]};
-    EXPECT_EQ(function.name(), orbit_name_mangled);
-    EXPECT_EQ(function.pretty_name(), orbit_name);
-    EXPECT_EQ(function.address(), 500);
-  }
-
-  // Now update the process modules in memory.
-  process.UpdateModuleInfos({module_info});
-  {
-    // Now ModuleManager should not be able to find the functions because the module is no longer in
-    // the memory.
-    std::vector<FunctionInfo> functions = module_manager.GetOrbitFunctionsOfProcess(process);
-    EXPECT_TRUE(functions.empty());
-  }
-}
-
 TEST(ModuleManager, GetModulesByFilename) {
   ModuleInfo module_info_1;
   module_info_1.set_name("name_1");
