+++ src/MetricsUploader/CaptureMetric.cpp
@@ -16,13 +16,6 @@ CaptureMetric::CaptureMetric(MetricsUploader* uploader, const CaptureStartData& 
   CHECK(uploader_ != nullptr);
   capture_data_.set_number_of_instrumented_functions(start_data.number_of_instrumented_functions);
   capture_data_.set_number_of_frame_tracks(start_data.number_of_frame_tracks);
-  capture_data_.set_number_of_manual_start_functions(start_data.number_of_manual_start_functions);
-  capture_data_.set_number_of_manual_stop_functions(start_data.number_of_manual_stop_functions);
-  capture_data_.set_number_of_manual_start_async_functions(
-      start_data.number_of_manual_start_async_functions);
-  capture_data_.set_number_of_manual_stop_async_functions(
-      start_data.number_of_manual_stop_async_functions);
-  capture_data_.set_number_of_manual_tracked_values(start_data.number_of_manual_tracked_values);
   capture_data_.set_thread_states(start_data.thread_states);
   capture_data_.set_memory_information_sampling_period_ms(
       start_data.memory_information_sampling_period_ms);
