+++ src/OrbitGl/TimeGraph.h
@@ -187,8 +187,6 @@ class TimeGraph : public orbit_gl::CaptureViewElement {
   [[nodiscard]] std::unique_ptr<orbit_accessibility::AccessibleInterface>
   CreateAccessibleInterface() override;
 
-  void ProcessOrbitFunctionTimer(orbit_client_protos::FunctionInfo::OrbitType type,
-                                 const orbit_client_protos::TimerInfo& timer_info);
   void ProcessApiEventTimer(const orbit_client_protos::TimerInfo& timer_info);
   void ProcessIntrospectionTimer(const orbit_client_protos::TimerInfo& timer_info);
   void ProcessValueTrackingTimer(const orbit_client_protos::TimerInfo& timer_info);
